relative extrusion for Marlin
